Deduplicate the provider code
