+++ apps/dav/lib/CalDAV/Activity/Provider/Calendar.php
@@ -24,13 +24,11 @@ namespace OCA\DAV\CalDAV\Activity\Provider;
 use OCP\Activity\IEvent;
 use OCP\Activity\IEventMerger;
 use OCP\Activity\IManager;
-use OCP\Activity\IProvider;
 use OCP\IL10N;
 use OCP\IURLGenerator;
-use OCP\IUser;
 use OCP\IUserManager;
 
-class Calendar implements IProvider {
+class Calendar extends Base {
 
 	const SUBJECT_ADD = 'calendar_add';
 	const SUBJECT_UPDATE = 'calendar_update';
@@ -49,14 +47,8 @@ class Calendar implements IProvider {
 	/** @var IManager */
 	protected $activityManager;
 
-	/** @var IUserManager */
-	protected $userManager;
-
 	/** @var IEventMerger */
 	protected $eventMerger;
-
-	/** @var string[] cached displayNames - key is the UID and value the displayname */
-	protected $displayNames = [];
 
 	/**
 	 * @param IL10N $l
@@ -66,10 +58,10 @@ class Calendar implements IProvider {
 	 * @param IEventMerger $eventMerger
 	 */
 	public function __construct(IL10N $l, IURLGenerator $url, IManager $activityManager, IUserManager $userManager, IEventMerger $eventMerger) {
+		parent::__construct($userManager);
 		$this->l = $l;
 		$this->url = $url;
 		$this->activityManager = $activityManager;
-		$this->userManager = $userManager;
 		$this->eventMerger = $eventMerger;
 	}
 
@@ -147,22 +139,6 @@ class Calendar implements IProvider {
 
 	/**
 	 * @param IEvent $event
-	 * @param string $subject
-	 * @param array $parameters
-	 */
-	protected function setSubjects(IEvent $event, $subject, array $parameters) {
-		$placeholders = $replacements = [];
-		foreach ($parameters as $placeholder => $parameter) {
-			$placeholders[] = '{' . $placeholder . '}';
-			$replacements[] = $parameter['name'];
-		}
-
-		$event->setParsedSubject(str_replace($placeholders, $replacements, $subject))
-			->setRichSubject($subject, $parameters);
-	}
-
-	/**
-	 * @param IEvent $event
 	 * @return array
 	 */
 	protected function getParameters(IEvent $event) {
@@ -212,59 +188,5 @@ class Calendar implements IProvider {
 		}
 
 		throw new \InvalidArgumentException();
-	}
-
-	/**
-	 * @param string $id
-	 * @return array
-	 */
-	protected function generateGroupParameter($id) {
-		return [
-			'type' => 'group',
-			'id' => $id,
-			'name' => $id,
-		];
-	}
-
-	/**
-	 * @param int $id
-	 * @param string $name
-	 * @return array
-	 */
-	protected function generateCalendarParameter($id, $name) {
-		return [
-			'type' => 'calendar',
-			'id' => $id,
-			'name' => $name,
-		];
-	}
-
-	/**
-	 * @param string $uid
-	 * @return array
-	 */
-	protected function generateUserParameter($uid) {
-		if (!isset($this->displayNames[$uid])) {
-			$this->displayNames[$uid] = $this->getDisplayName($uid);
-		}
-
-		return [
-			'type' => 'user',
-			'id' => $uid,
-			'name' => $this->displayNames[$uid],
-		];
-	}
-
-	/**
-	 * @param string $uid
-	 * @return string
-	 */
-	protected function getDisplayName($uid) {
-		$user = $this->userManager->get($uid);
-		if ($user instanceof IUser) {
-			return $user->getDisplayName();
-		} else {
-			return $uid;
-		}
 	}
 }
